Remove 'translated' XLIFF attribute
Remove all references to the 'translated' attribute on the <trans-unit>
tag. This attribute is not valid XLIFF and is ignored by Weblate.

Weblate now manages this using the 'state' attribute on the <target>
tag. Managing the 'state' will be transparent to AtoM. If the 'state'
is not present, Weblate uses the presence of target string to determine
the current state.